Split out LeetCode helpers into their own package
Also a chance to play with [Workspaces](https://classic.yarnpkg.com/lang/en/docs/workspaces/#toc-how-to-use-it).
