+++ tools/leetcode-api/src/sleep.ts
@@ -1,0 +1,5 @@
+export function sleep(ms: number): Promise<void> {
+  return new Promise((resolve) => {
+    setTimeout(resolve, ms);
+  });
+}
