+++ tools/post-potd/.gitignore
@@ -1,4 +1,1 @@
-node_modules/
-yarn-error.log
-
 secrets_DO_NOT_COMMIT_OR_SHARE.json
